Add following? method that doesn't require user
When all I want to do is check if a user is following a given channel, having the overhead of fetching a second unnecessary API request may make nice Ruby semantic sense but doesn't make great performance sense to me.

Add this method that accepts a name and channel name and fires a single API request with that information, cutting out the extraneous step.